requirements.txt
beautifulsoup4==4.10.0
Brotli==1.0.9
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
dash==2.1.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
distlib==0.3.4
filelock==3.4.2
Flask==2.0.2
Flask-Compress==1.10.1
free-proxy==1.0.6
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
Js2Py==0.71
lxml==4.7.1
MarkupSafe==2.0.1
multitasking==0.0.10
numpy==1.22.2
pandas==1.3.5
pandas-ta==0.3.14b0
platformdirs==2.4.1
plotly==5.6.0
pyfinviz==0.7
pyjsparser==2.7.1
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
tenacity==8.0.1
tqdm==4.62.3
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
validators==0.18.2
virtualenv==20.13.0
Werkzeug==2.0.3
yfinance==0.1.69